+++ app/src/main/java/com/byeboo/app/presentation/quest/QuestState.kt
@@ -10,9 +10,9 @@ sealed class QuestState {
 }
 
 sealed interface QuestSideEffect {
-    data class NavigateToQuestTip(val questId: Int) : QuestSideEffect
-    data class NavigateToQuestRecording(val questId: Int) : QuestSideEffect
-    data class NavigateToQuestBehavior(val questId: Int) : QuestSideEffect
-    data class NavigateToQuestReview(val questId: Int, val questType: QuestType) : QuestSideEffect
+    data class NavigateToQuestTip(val questId: Long) : QuestSideEffect
+    data class NavigateToQuestRecording(val questId: Long) : QuestSideEffect
+    data class NavigateToQuestBehavior(val questId: Long) : QuestSideEffect
+    data class NavigateToQuestReview(val questId: Long, val questType: QuestType) : QuestSideEffect
     data object NavigateToHome : QuestSideEffect
 }
