+++ app/src/main/java/com/byeboo/app/presentation/quest/behavior/QuestBehaviorState.kt
@@ -6,7 +6,7 @@ import com.byeboo.app.domain.model.QuestWritingState
 data class QuestBehaviorState(
     val stepNumber: Int = 1,
     val stepMissionTitle: String? = null,
-    val questId: Int = 1,
+    val questId: Long = 1,
     val questNumber: Int = 1,
     val questTitle: String = "",
     val imageCount: Int = 0,
@@ -19,6 +19,6 @@ data class QuestBehaviorState(
 
 sealed interface QuestBehaviorSideEffect {
     data object NavigateToQuest : QuestBehaviorSideEffect
-    data class NavigateToQuestTip(val questId: Int) : QuestBehaviorSideEffect
-    data class NavigateToQuestBehaviorComplete(val questID: Int) : QuestBehaviorSideEffect
+    data class NavigateToQuestTip(val questId: Long) : QuestBehaviorSideEffect
+    data class NavigateToQuestBehaviorComplete(val questID: Long) : QuestBehaviorSideEffect
 }
