+++ app/src/main/java/com/byeboo/app/presentation/quest/behavior/navigation/QuestBehaviorRoute.kt
@@ -6,8 +6,8 @@ import kotlinx.serialization.Serializable
 @Serializable
 sealed class QuestBehavior : Route {
     @Serializable
-    data class QuestBehaviorWriting(val questId: Int) : QuestBehavior()
+    data class QuestBehaviorWriting(val questId: Long) : QuestBehavior()
 
     @Serializable
-    data class QuestBehaviorComplete(val questId: Int) : QuestBehavior()
+    data class QuestBehaviorComplete(val questId: Long) : QuestBehavior()
 }
