+++ app/src/main/java/com/byeboo/app/presentation/quest/model/QuestModel.kt
@@ -10,7 +10,7 @@ data class QuestGroup(
 )
 
 data class Quest(
-    val questId: Int = 0,
+    val questId: Long = 0,
     val questNumber: Int = 0,
     val questQuestion: String = "연애에서 반복됐던 문제 패턴 3가지를 생각해보아요.",
     val state: QuestState = QuestState.Available,
